+++ fst/main.py
@@ -1,11 +1,10 @@
-import os
 import click
 
 from fst.file_utils import get_models_directory
 from fst.query_handler import handle_query, DynamicQueryHandler
 from fst.directory_watcher import watch_directory
 from fst.logger import setup_logger
-from fst.constants import CURRENT_WORKING_DIR
+from fst.config_defaults import CURRENT_WORKING_DIR
 
 
 @click.group()
